+++ src/Contract/Provider/Provider.php
@@ -18,7 +18,7 @@ interface Provider
      *
      * @param \Zapheus\Contract\Container\Writable $container
      *
-     * @return \Zapheus\Contract\Container\Container
+     * @return \Zapheus\Contract\Container\Writable
      */
     public function register(Writable $container);
 }