+++ src/Http/ServerProvider.php
@@ -35,7 +35,7 @@ class ServerProvider implements Provider
      *
      * @param \Zapheus\Contract\Container\Writable $container
      *
-     * @return \Zapheus\Contract\Container\Container
+     * @return \Zapheus\Contract\Container\Writable
      */
     public function register(Writable $container)
     {