+++ src/Renderer/RendererProvider.php
@@ -34,7 +34,7 @@ class RendererProvider implements Provider
      *
      * @param \Zapheus\Contract\Container\Writable $container
      *
-     * @return \Zapheus\Contract\Container\Container
+     * @return \Zapheus\Contract\Container\Writable
      */
     public function register(Writable $container)
     {