+++ tests/Fixture/Providers/TestProvider.php
@@ -15,7 +15,7 @@ class TestProvider implements Provider
     /**
      * @param \Zapheus\Contract\Container\Writable $container
      *
-     * @return \Zapheus\Contract\Container\Container
+     * @return \Zapheus\Contract\Container\Writable
      */
     public function register(Writable $container)
     {