+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to `Zapheus` will be documented in this file.
 
 ## 0.2.0 - Unreleased
 
+### Added
+- `getContainer` in `Application`
+
 ### Changed
 - Moved interfaces to `Zapheus\Contract`